test(macos): name and focus host platform tests
